requirements.txt
flask
gunicorn
tensorflow-cpu
numpy
opencv-python-headless
gTTS
requests
reportlab
python-dotenv
google-generativeai
